Kiwiplan : Module PaperRolll Done.

+++ database/migrations/2017_09_07_135913_create_purchase_orders_table.php
@@ -22,7 +22,7 @@ class CreatePurchaseOrdersTable extends Migration
           $table->string('po_num');
           $table->string('po_num_ex')->nullable();
           $table->datetime('po_date');
-          $table->decimal('po_qty',7,2);
+          $table->float('po_qty');
           $table->string('due_date');
           $table->string('contact_person')->nullable();
           $table->integer('term');
